# expr: Optimizing for integer values