--- a/src/uu/expr/src/expr.rs
+++ b/src/uu/expr/src/expr.rs
@@ -13,7 +13,7 @@ use uucore::{
     format_usage, help_about, help_section, help_usage,
 };
 
-use crate::syntax_tree::{is_truthy, NumOrStr};
+use crate::syntax_tree::is_truthy;
 
 mod syntax_tree;
 
@@ -108,9 +108,9 @@ pub fn uumain(args: impl uucore::Args) -> UResult<()> {
         .map(|v| v.into_iter().map(|s| s.as_ref()).collect::<Vec<_>>())
         .unwrap_or_default();
 
-    let res: String = AstNode::parse(&token_strings)?.eval()?.into();
+    let res: String = AstNode::parse(&token_strings)?.eval()?.eval_as_string();
     println!("{res}");
-    if !is_truthy(&NumOrStr::from(res)) {
+    if !is_truthy(&res.into()) {
         return Err(1.into());
     }
     Ok(())