An attempt at resolving issue https://github.com/uutils/coreutils/issues/5591