@sylvestre 2023-12-03
nice, did you do some benchmark ? - just curious on what kind of impact this can have
(hyperfine is pretty good for this)