@Arp-1 2023-12-03
Not yet, was looking if there are some open test data sets available for `expr`. Will look into `hyperfine`, thanks for your suggestion!