@Arp-1 2023-12-09
Sorry for getting back on this so late, had a busy week!
@tertsdiepraam thanks for your review, have made fixes for all your comments.
I did some quick & dirty benchmarking on my local system (terrible reliability, got inaccuracies of ~30%) with `hyperfine`. For small numerical expressions( < 10 operations), I saw almost no gains (5% at best). For bigger numerical expressions (~100 operations) with huge numbers (< 10^150), saw an improvement of around 40%. For string operations, There was a performance loss of no more than 5%.
Tested release build of `expr` with `shell=none` - main branch (https://github.com/uutils/coreutils/commit/8b1a82edb9a1f47652ac55a536066ade3b28534e) against my branch (https://github.com/Arp-1/coreutils/commit/4d2ae8485cd65429e64606119acac3156158ea2b)
Let me know if I should put more details of my testing setup.