@Arp-1 2023-12-10
Thanks for taking a look at it again @tertsdiepraam !
On your benchmarking question, yes, I'd built `expr` with `--release`. Though reiterating on the instability of the runtimes on my local system, the spread of time taken for optimized `expr` was 0.74ms to 2.23ms and for unoptimized `expr` it was 0.73 ms to 2.85ms:

![expr_str](https://github.com/uutils/coreutils/assets/37845584/0da45d8d-bcaa-46a4-8e33-1d6bfd892dcb)

Maybe I should try and do this on a little less noisy system :)